requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
psycopg2-binary
python-jose[cryptography]
passlib[argon2]
argon2-cffi
pydantic
openai
email-validator

